requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
